--- a/pom.xml
+++ b/pom.xml
@@ -31,6 +31,8 @@ under the License.
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<flink.version>1.11-SNAPSHOT</flink.version>
+		<flink.shaded.version>10.0</flink.shaded.version>
+		<netty.tcnative.version>2.0.25.Final</netty.tcnative.version>
 		<java.version>1.8</java.version>
 		<scala.binary.version>2.11</scala.binary.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
@@ -396,6 +398,44 @@ under the License.
 					</plugin>
 				</plugins>
 			</build>
+		</profile>
+
+		<!--
+            This module is not provided by Apache Flink by default due to licensing issues
+            https://issues.apache.org/jira/browse/LEGAL-393
+            It can be used instead of relying on dynamically-linked OpenSSL libs by building
+            it manually in flink-shaded:
+            > mvn clean install -Pinclude-netty-tcnative-static -pl flink-shaded-netty-tcnative-static
+        -->
+		<profile>
+			<id>include-netty-tcnative-static</id>
+			<activation>
+				<property>
+					<name>include-netty-tcnative-static</name>
+				</property>
+			</activation>
+			<dependencies>
+				<dependency>
+					<groupId>org.apache.flink</groupId>
+					<artifactId>flink-shaded-netty-tcnative-static</artifactId>
+					<version>${netty.tcnative.version}-${flink.shaded.version}</version>
+				</dependency>
+			</dependencies>
+		</profile>
+		<profile>
+			<id>include-netty-tcnative-dynamic</id>
+			<activation>
+				<property>
+					<name>!include-netty-tcnative-static</name>
+				</property>
+			</activation>
+			<dependencies>
+				<dependency>
+					<groupId>org.apache.flink</groupId>
+					<artifactId>flink-shaded-netty-tcnative-dynamic</artifactId>
+					<version>${netty.tcnative.version}-${flink.shaded.version}</version>
+				</dependency>
+			</dependencies>
 		</profile>
 	</profiles>