This PR and adds OpenSSL network benchmarks:

These can be executed either with dynamically-linked OpenSSL libraries or
statically-linked (and manually-provided) ones. Instructions are in README.md.

Current results on my PC:
- JDK 8:
```
Benchmark                                                   (channelsFlushTimeout)   Mode  Cnt      Score     Error   Units
StreamNetworkThroughputBenchmarkExecutor.networkThroughput          1000,100ms,SSL  thrpt   30  11760.342 ± 282.592  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput      1000,100ms,OpenSSL  thrpt   30  27889.053 ± 893.833  ops/ms
```
- JDK11:
```
Benchmark                                                   (channelsFlushTimeout)   Mode  Cnt      Score     Error   Units
StreamNetworkThroughputBenchmarkExecutor.networkThroughput          1000,100ms,SSL  thrpt   30  20628.476 ± 628.461  ops/ms
StreamNetworkThroughputBenchmarkExecutor.networkThroughput      1000,100ms,OpenSSL  thrpt   30  27289.926 ± 962.960  ops/ms
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dataartisans/flink-benchmarks/56)
<!-- Reviewable:end -->