@Myasuka 2020-04-30
Just remind whether this PR need to be merged before [FLINK-17281](https://issues.apache.org/jira/browse/FLINK-17281) is resolved.